+++ src/Render/interactions/index.ts
@@ -3,3 +3,4 @@ export { enableZoom } from "./zoom";
 export { enableReferenceLine } from "./referenceLine";
 export { enableDragOutside } from "./dragOutside";
 export { enableSelection } from "./selection";
+export { enableKeyMove } from "./keyMove";