+++ src/Render/types/tools.ts
@@ -10,4 +10,6 @@ export interface SelectionTool {
 	select: (nodes: Konva.Node[]) => void;
 	/** 清除所有选择 */
 	selectingClear: (slient?: boolean) => void;
+	/** 通过偏移量移动选中的节点 */
+	selectingNodesMove: (offset: Konva.Vector2d) => void;
 }